% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{appendInfo}
\alias{appendInfo}
\alias{appendInfo,info-method}
\title{append width and step size to an \code{info} object}
\usage{
appendInfo(obj, w, s)

\S4method{appendInfo}{info}(obj, w, s)
}
\arguments{
\item{obj}{An \code{info} object.}

\item{w}{The flanking window size in basepairs (integer).}

\item{s}{The step size for coverage value sampled (integer).}
}
\value{
An updated \code{info} object
}
