% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{avgplot}
\alias{avgplot}
\alias{avgplot,matlist,chip,info-method}
\title{Plot average coverage signals stratified by clusters and by individual/groupled samples}
\usage{
avgplot(matlist.obj, chip.obj, info.obj, pngfoutFe, pngfoutSm)

\S4method{avgplot}{matlist,chip,info}(matlist.obj, chip.obj, info.obj,
  pngfoutFe, pngfoutSm)
}
\arguments{
\item{matlist.obj}{A \code{matlist} object}

\item{chip.obj}{A \code{chip} object}

\item{info.obj}{An \code{info} object}

\item{pngfoutFe}{An png graph of average plot stratified by clusters}

\item{pngfoutSm}{An png graph of average plot stratified by individual/groupled samples}
}
\value{
No object
}
\description{
Plot average coverage signals stratified by clusters and by individual/groupled samples.
}
