% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgClass.R
\docType{class}
\name{chip-class}
\alias{chip-class}
\alias{chip}
\title{An S4 class to represent bed file of chip-seq peaks.}
\description{
Store cluster sets in a bed object.
}
\section{Slots}{

\describe{
\item{\code{bed}}{data.frame object}
}}

