% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgConst.R
\name{chipConst}
\alias{chipConst}
\title{construct an \code{chip} object}
\usage{
chipConst(chipF)
}
\arguments{
\item{chipF}{a ChIP-seq bed file}
}
\value{
A \code{chip} object
}
\description{
construct an \code{chip} object from args$chip
}
