% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{clusterR}
\alias{clusterR}
\alias{clusterR,chip,info,matlist-method}
\title{Kmeans subclustering within each clusters of peaks}
\usage{
clusterR(chip.obj, info.obj, matlist.obj, nms)

\S4method{clusterR}{chip,info,matlist}(chip.obj, info.obj, matlist.obj, nms)
}
\arguments{
\item{chip.obj}{A \code{chip} object.}

\item{info.obj}{An \code{info} object.}

\item{matlist.obj}{A \code{matlist} object.}

\item{nms}{A character vector of length one, specifying the directory and the prefix for plotting the optimum number of subclusters within each cluster.}
}
\value{
An updated \code{chip} object.
}
\description{
Find the optimum number of subclusters within each clusters,
Kmeans sub-clustering within each clusters,
Update the cluster labels (the 4th column) of the \code{bed} slot of a \code{chip} object,
by labels of the combinations of cluster and subcluster labels.
}
