% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgClass.R
\docType{class}
\name{matlist-class}
\alias{matlist-class}
\alias{matlist}
\title{An S4 class to represent coverage value and ylab for plotting}
\description{
Store
coverage value of all samples in a list (list ll).
ylal for plotting.
}
\section{Slots}{

\describe{
\item{\code{ll}}{list object representing a list of samples' coverage value sampled.}

\item{\code{ylab}}{expression object for plotting.}
}}

