% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgConst.R
\name{matlistConst}
\alias{matlistConst}
\title{construct an \code{matlist} object}
\usage{
matlistConst(config)
}
\arguments{
\item{config}{A BW.config file}
}
\value{
A \code{matlist} object
}
\description{
construct an \code{matlist} object according to a BW.config file.
}
