% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{mergeRep}
\alias{mergeRep}
\alias{mergeRep,matlist-method}
\title{Merge within-group smaples' coverage signals}
\usage{
mergeRep(matlist.obj, repconfig)

\S4method{mergeRep}{matlist}(matlist.obj, repconfig)
}
\arguments{
\item{matlist.obj}{A \code{matlist} object.}

\item{repconfig}{A replicates configuration file (1st column: individual sample names as in the 2nd column of BW.config; 2nd column: group names; columns separated by tab).}
}
\value{
A sample-grouped \code{matlist} object
}
\description{
Merge within-group smaples' coverage signals, based on the specification in the repconfig file.
}
