% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterR_Optimal_Clusters_KMeans.R
\name{Optimal_Clusters_KM}
\alias{Optimal_Clusters_KM}
\title{Find the optimal number of kmeans clusters}
\usage{
Optimal_Clusters_KM(data, max_clusters, criterion = "variance_explained",
  fK_threshold = 0.85, num_init = 1, max_iters = 200,
  initializer = "optimal_init", threads = 1, tol = 1e-04,
  plot_clusters = TRUE, verbose = FALSE, tol_optimal_init = 0.3,
  seed = 1)
}
\arguments{
\item{data}{matrix or data.frame object}

\item{max_clusters}{numeric/integer vector of length 1}
}
\value{
cluster membership integer vector
}
\description{
Find the optimal number of kmeans clusters
by supplying data (matrix or data.frame), and
specifying max_clusters (the maximum number of clusters).
The function plots the optimal cluster number plot and returns cluster membership (integers).
}
