% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{orderbyIntIdx}
\alias{orderbyIntIdx}
\alias{orderbyIntIdx,info-method}
\title{orderbyIntIdx}
\usage{
orderbyIntIdx(obj, IntIdx)

\S4method{orderbyIntIdx}{info}(obj, IntIdx)
}
\arguments{
\item{obj}{An \code{info} object}

\item{IntIdx}{An integer vector of the same length as the currently kept peaks.}
}
\value{
An updated \code{info} object
}
\description{
Order peaks by currently given integer index,
which corresponds to the currently sub-indices of currently kept peaks.
}
