% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{orderClus}
\alias{orderClus}
\alias{orderClus,chip-method}
\title{order clusters in the 4th col of a bed file according to a configuration file}
\usage{
orderClus(obj, orderconfig)

\S4method{orderClus}{chip}(obj, orderconfig)
}
\arguments{
\item{obj}{A \code{chip} object.}

\item{orderconfig}{An order configruation file (character).}
}
\value{
A list of an level-order changed \code{chip} object and an \code{info} object.
}
\description{
order clusters in the 4th col of bed file based on a configuration file with 1 column of ordered clusters.
}
