% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{outliers}
\alias{outliers}
\alias{outliers,info,chip,matlist-method}
\title{set currently kept peaks to be discarded if any NA coverage value is reported in any individual samples.}
\usage{
outliers(info.obj, chip.obj, matlist.obj)

\S4method{outliers}{info,chip,matlist}(info.obj, chip.obj, matlist.obj)
}
\arguments{
\item{info.obj}{An \code{info} object}

\item{chip.obj}{A \code{chip} object}

\item{matlist.obj}{A \code{matlist} object}
}
\value{
An updated \code{info} object
}
\description{
Set
Currently kept peaks (kpt.idx) to FALSE,
had coverage values for a peak reported to be NA in any samples.
}
