% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{reord}
\alias{reord}
\alias{reord,info,chip,matlist-method}
\title{Reorder peaks (specified info.obj) by the descending order of coverage signals of individual/grouped samples.}
\usage{
reord(info.obj, chip.obj, matlist.obj)

\S4method{reord}{info,chip,matlist}(info.obj, chip.obj, matlist.obj)
}
\arguments{
\item{info.obj}{An \code{info} object}

\item{chip.obj}{A \code{chip} object}

\item{matlist.obj}{A \code{matlist} object}
}
\value{
An updated \code{info} object
}
\description{
Reorder peaks (specified info.obj) by the descending order of coverage signals of individual/grouped samples.
Update info.obj by the new peak order.
}
