% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{returnIntIdx}
\alias{returnIntIdx}
\alias{returnIntIdx,info-method}
\title{integer index (peak order) to keep of an \code{info} object}
\usage{
returnIntIdx(obj)

\S4method{returnIntIdx}{info}(obj)
}
\arguments{
\item{obj}{An \code{info} object.}
}
\value{
An integer index for ordered peaks to keep
}
\description{
Return integer index (peak order) that is supposed to be kept
}
