% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{rmSmallClus}
\alias{rmSmallClus}
\alias{rmSmallClus,chip,info-method}
\title{remove small (size<1000) clusters}
\usage{
rmSmallClus(info.obj, chip.obj)

\S4method{rmSmallClus}{chip,info}(info.obj, chip.obj)
}
\arguments{
\item{info.obj}{An \code{info} object}

\item{chip.obj}{A \code{chip} object}
}
\value{
An updated \code{info} object
}
\description{
Set
Currently kept peaks (kpt.idx) to FALSE,
had the size of cluster is less than 1000.
}
