% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{subsetbyBoolIdx}
\alias{subsetbyBoolIdx}
\alias{subsetbyBoolIdx,info-method}
\title{subset by boolean index}
\usage{
subsetbyBoolIdx(obj, BoolIdx)

\S4method{subsetbyBoolIdx}{info}(obj, BoolIdx)
}
\arguments{
\item{obj}{An \code{info} object}

\item{BoolIdx}{A boolean vector of the length as the currently kpt peaks}
}
\value{
An updated \code{info} object
}
\description{
Filter peaks to keep or discard.
Set certain TRUE values of kpt.idx of an \code{info} object to FALSE
identified by the boolean values specified in BoolIdx.
TRUE values in BoolIdx specify of those currently kept peaks to be kept downstream.
FLASE values in BoolIdx specify of those currently kept peaks to be filtered out downstream.
}
