% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{subsetbyIntIdx}
\alias{subsetbyIntIdx}
\alias{subsetbyIntIdx,info-method}
\title{subset by integer index}
\usage{
subsetbyIntIdx(obj, IntIdx)

\S4method{subsetbyIntIdx}{info}(obj, IntIdx)
}
\arguments{
\item{obj}{An \code{info} object}

\item{IntIdx}{An integer/numeric vector specifying index of currently kept peaks to be filtered out from the current point.}
}
\value{
An updated \code{info} object
}
\description{
Filter peaks to keep or discard.
Set certain TRUE values of kpt.idx of an \code{info} object to FALSE
identified by the integer specified in IntIdx.
The function raises an error,
if the length of index vector is longer or equal to the number of peaks to be kpt.
The function also raises an error,
if the max integer position is larger than the lagest indexable postion of peaks to be kpt.
}
