% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{tdplot}
\alias{tdplot}
\alias{tdplot,matlist,chip,info-method}
\title{Plot a tornado plot}
\usage{
tdplot(matlist.obj, chip.obj, info.obj, pdffout)

\S4method{tdplot}{matlist,chip,info}(matlist.obj, chip.obj, info.obj, pdffout)
}
\arguments{
\item{matlist.obj}{A \code{matlist} object}

\item{chip.obj}{A \code{chip} object}

\item{info.obj}{An \code{info} object}

\item{pdffout}{An output pdf file for the tornado plot}
}
\value{
No object
}
\description{
Plot a tornado plot
}
