% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{updateChip}
\alias{updateChip}
\alias{updateChip,chip,info-method}
\title{Update the \code{bed} slot of the \code{chip} object by the order and filteration specified in the \code{info} object}
\usage{
updateChip(chip.obj, info.obj)

\S4method{updateChip}{chip,info}(chip.obj, info.obj)
}
\arguments{
\item{chip.obj}{A \code{chip} object}

\item{info.obj}{An \code{info} object}
}
\value{
An updated \code{chip} object
}
\description{
Update the \code{bed} slot of the \code{chip} object,
by the order and filteration specified in the \code{info} object.
Once updated, the kpt.idx and ord.idx slots of the \code{info} object become obselete for the \code{chip} object.
}
