% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{updateClus}
\alias{updateClus}
\alias{updateClus,chip-method}
\title{Update the label of peaks in \code{bed} slot of an \code{chip} object}
\usage{
updateClus(chip.obj, IntIdx, subclus)

\S4method{updateClus}{chip}(chip.obj, IntIdx, subclus)
}
\arguments{
\item{chip.obj}{A \code{chip} object.}

\item{IntIdx}{An integer/numeric vector specify a subset or all of labels in the \code{bed} slot of \code{chip} object to be updated by given factor/character vector of subclus.}

\item{subclus}{A factor/character vector of the same length as IntIdx specifying new labels for peak clusters.}
}
\value{
An updated \code{chip} object
}
\description{
Update
the label of peaks (the 4th column) in \code{bed} slot of an \code{chip} object,
according to the integer position (IntIdx),
by the new lables (subclus).
}
