% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{updateMatlist}
\alias{updateMatlist}
\alias{updateMatlist,matlist,info-method}
\title{Update the \code{ll} slot of the \code{matlist} object by the order and filteration specified in the \code{info} object}
\usage{
updateMatlist(matlist.obj, info.obj)

\S4method{updateMatlist}{matlist,info}(matlist.obj, info.obj)
}
\arguments{
\item{matlist.obj}{A \code{matlist} object}

\item{info.obj}{An \code{info} object}
}
\value{
An updated \code{matlist} object
}
\description{
Update the \code{ll} slot of the \code{mat} object,
by the order and filteration specified in the \code{info} object.
Once updated, the kpt.idx and ord.idx slots of the \code{info} object become obselete for the \code{matlist} object.
}
