% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{writeChip}
\alias{writeChip}
\alias{writeChip,chip-method}
\title{Output the \code{bed} slot of the \code{chip} object to file.}
\usage{
writeChip(chip.obj, fout)

\S4method{writeChip}{chip}(chip.obj, fout)
}
\arguments{
\item{chip.obj}{A \code{chip} object}

\item{fout}{A output text file}
}
\value{
invisible of NULL
}
\description{
Output the \code{bed} slot of the \code{chip} object to file.
}
