% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{bplot}
\alias{bplot}
\alias{bplot,matlist,chip,info-method}
\title{Plot a box plot}
\usage{
bplot(matlist.obj, chip.obj, info.obj, pdffoutFe, pdffoutSm)

\S4method{bplot}{matlist,chip,info}(matlist.obj, chip.obj, info.obj, pdffoutFe,
  pdffoutSm)
}
\arguments{
\item{matlist.obj}{A \code{matlist} object}

\item{chip.obj}{A \code{chip} object}

\item{info.obj}{An \code{info} object}

\item{pdffoutFe}{A pdf graph of box/violin plot stratified by clusters}

\item{pdffoutSm}{A pdf graph of box/violin plot stratified by individual/groupled samples}
}
\value{
No object
}
\description{
Plot a box plot
}
