% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TdAvgGeneric.R, R/TdAvgMethods.R
\docType{methods}
\name{subsetbyIntIdx}
\alias{subsetbyIntIdx}
\alias{subsetbyIntIdx,info,numeric,logical-method}
\title{subset by integer index}
\usage{
subsetbyIntIdx(obj, IntIdx, invert)

\S4method{subsetbyIntIdx}{info,numeric,logical}(obj, IntIdx, invert)
}
\arguments{
\item{obj}{An \code{info} object}

\item{IntIdx}{An integer/numeric vector specifying index of currently kept peaks to be filtered out from the current point.}

\item{invert}{when TRUE, specifies selecting the rows, or otherwise FALSE, specifies the complement rows.}
}
\value{
An updated \code{info} object
}
\description{
Filter peaks to keep or discard.
Order of full/subset of original bed rows specified in the integer specified in IntIdx.
The function raises an error,
if the length of index vector is longer or equal to the number of peaks to be kpt.
The function also raises an error,
if the max integer position is larger than the lagest indexable postion of peaks to be kpt.
}
